+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Upgraded to Unity 2019. Dropped support for Unity 2018.
 - Removed `dynamicBatching` from Render Pipeline Asset. Configure dynamic batching in `Project Settings/Player/Others Settings` instead.
 
+### Added
+
+- Added indirect lighting modifier.
+
 ### Fixed
 
 - Fixed the gizmos being drawn in Game View.